requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
django-environ==0.10.0 ; python_version >= "3.11" and python_version < "4"
django-extensions==3.2.3 ; python_version >= "3.11" and python_version < "4.0"
django-rest-framework==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
django-tastypie==0.14.5 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.4 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.11" and python_version < "4.0"
python-mimeparse==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
